fix: Make node apply a default store path (#4186)
## Relevant issue(s)

Resolves #4184

## Description

If a Defra node was created without the `node.WithStorePath` option, the
default store path, `$HOME/.defradb`, would never actually get set. The
`cli` package did contain a function, `setContextRootDir` which set such
a path in its context, but this logic was unique to that package. This
PR changes this by creating a new function in the `node` package,
`GetDefaultStorePath`. This function is used by the
`DefaultStoreOptions` function, to apply the default value to the `path`
field. Additionally, this function is called by the CLI inside the
`setContextRootDir` function so as not to break any of the ways context
is passed around or used, while also making sure the logic remains
consistent across clients, without ever being repeated in different
locations.

One note, is that this PR does not yet adjust how the way in which the C
bindings package pass in the `node.WithStorePath` option to the new
node. The reason for this, is that a PR already exists for exactly that
issue. Presumably, it can be closed and merged after this one gets
merged and closed.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

I used the C bindings on WSL and on Android, and the CLI with WSL.

Specify the platform(s) on which this was tested:
- WSL
- Android