chore: Utilize default db path if blank string passed in (#4173)
## Relevant issue(s)

Resolves #4160 

## Description

Previously, the `CreateNode` function of the c bindings would return an
error if the `inMemory` flag was set to 0, but the `dbPath` field was a
blank string. This PR changes the behavior to rely on Go's
`os.UserHomeDir()` function to create a `.defradb` directory in the home
directory, instead of erroring out.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

(*replace*) Describe the tests performed to verify the changes. Provide
instructions to reproduce them.

Specify the platform(s) on which this was tested:
- WSL
- Android