fix: Introspection error and badly formatted GQL types (#4175)
## Relevant issue(s)

Resolves #4132 

## Description

Fixes the introspection issue. Theres room for discussion with the
solution. Specifically the new `ScalarAggregateNumericBlock` type. Which
includes a "noop" field `_`, since the `gql.InputObject` type requires
at least one field to be defined.

This wasn't needed before, but the method to get around this, also
causes the introspection error in the first place. There might be away
around it in the depths of the gql library but its not self evident how.

In reality, this is a symptom of *how* we target/select fields for
aggregates, and not a issue with the gql lib itself.

This PR also introduces a new CI test workflow to handle using nodejs in
a specific unit test. We need to test the actual GraphQL client that
most JS based client use to verify introspection results, which means we
need to use the actual library via node. There is a new test in the
`internal/request/graphql/schema` package specifically verifying the
introspection output. Note this is a unit test, and therefore obviously
doesn't cover some of the code paths between an external introspection
request finding its way all the way to the gql lib, just the important
90% of it. I'm open if people have strong feelings to include a full
integration test for this.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [ x I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Updated existing integration tests. Added a gql introspection
conformance test.

Specify the platform(s) on which this was tested:
- Ubuntu (WSL2)