refactor: KMS sync ordering (#4214)
## Relevant issue(s)

Resolves #4213

## Description

This PR makes a change to the p2p and KMS sync process that fixes an
issue I ran into when working on the LevelDB store. It is also
organizationally a little better since all of the p2p fetching logic now
resides in the p2p package.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Integration tests

Specify the platform(s) on which this was tested:
- MacOS