feat: Add ability to bypass DAC for performance (#4115)
## Relevant issue(s)
Resolves #3886

## Description
- Implementation of DAC bypass ~~using a cache to store~~ if an identity
has the ability to dac bypass ~~or not~~.
- There are a lot of tests because in previous work I skipped testing
DAC access in combination with NAC. I have tested most if not all edge
case concerns for DAC (public-doc, private-owned doc, private-not-owned
own) <> NAC (on and temporarily turned off) <> Relationship (granting,
revoking/not having)
- It is a bit odd to 100% assert DAC-bypass happens in the tests, as we
are seeing "behavior" of dac-bypass working, not asserting it actually
occurred (perhaps can find a better way to assert that in future).

### Concerns / Discussion
I do have some concerns with the caching approach:
1) local acp calls are fast anyways so can get rid of caching and just
make per request call to acp.
2) the caching approach will need to be scrapped if/when we implement
sourcehub/global nac.
3) maintaining the cache state can be messy (atm we are clearing it
through relationship add/delete calls so it's not too bad).
4) seems like we are holding/replicating the acp state that is part of
the acp systems job perhaps not the application's.

However despite above, I am happy to merge this as is, because it works
for what we have now (local NAC) and the testing covers the concerns I
value the most.

#### EDIT: Removed cache completely according the standup discussion, so
above is irrelevant now

## Todo
- [x] Removed cache completely according the standup discussion
- [x] Rename `db/permission` package to `db/acp`