chore: Make C bindings ViewRefresh take C.CollectionOptions (#4239)
## Relevant issue(s)

Resolves #4204

## Description

This changes the function signature of the `ViewRefresh` function of the
C bindings. It now uses a `C.CollectionOptions` struct instead of
individual parameters.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL