feat: Allow C binding view functions to take in identities (#4241)
## Relevant issue(s)

Resolves #4240 

## Description

Moving forward, we will be introducing gated permissions to the
`ViewAdd` and `ViewRefresh` functions. When that happens, we would
expect the integration tests to fail for the C client, and for the
functionality to not be accessible to users of the C bindings. The C
bindings, before this PR, just did not take in an identity parameter for
these two functions. This PR is to get ahead of that problem before it
ever crops up, by allowing these functions to take in an identity
parameter.

Appropriate changes have been made to the exported functions, as well as
to the wrapper functions.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL