feat: Upgrade CommitLink with full Commit object support (#4217)
## Relevant issue(s)

Resolves #4149 

## Description

This replaces the current `CommitLink` object in the `links` field of a
`Commit` to use a recursive nested `Commit` type for the `links.

It also breaks up the `links` field into two fields, `links` and
`heads`, which contain the `block.Links` and `block.Heads` respectively.
A consequence of this is the removal of the `name` field in the previous
`links`, which can be replaced with the original `Commit` type
`fieldName` field.

Almost all of the tests that originally used `links` has been migrated
to use both `links` and `heads` (even in some cases where the `heads`
will be empty for that particular test). Only a few haven't been
changed. Don't read into which specific tests, it was somewhat random by
the end.

One small consequence of this split of `links` and `heads` is how
`_count` within a `Commit` object works. Specifically, no longer being
able to count *all* the links as we did before, and would have to issue
two `_count(field: links)` and `_count(field: heads)`. Additionally, the
`_count` for `links`/`heads` is limited in functionality to a single
base field, instead of being able to do a more complex sub selection on
the underlying `Commit` type. I believe this to a perfectly fine design
approach *for now*.

## Tasks
- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?
Extended integration tests to cover the new behavior, as well as notable
refactor of existing integration tests.

Specify the platform(s) on which this was tested:
- Ubuntu (WSL2)