refactor: Wrap Datastore in key-typed wrapper (#4242)
## Relevant issue(s)

Resolves #4233

## Description

Wraps the Datastore in a new `Keyedstore` wrapper around
`corekv.ReaderWriter`.

This gives us a single location from which we can manage Defra-specific
access to the datastore using they typed keys, which for #4207, will
mean enforcing read-locks for collection-specific keys whilst a truncate
is being performed.

The new `Keyedstore` interface is store agnostic and can be used for our
other stores, and may need to be at some point. I have only used it for
the Datastore in this PR, as I think that is the only store that I will
need/want to lock in such a way for `truncate`.

The only implementation of `Keyedstore` would work for for all stores at
the moment, however that will change soon, within the `truncate` PR, as
the locking will be very specific to the datastore.