fix: Make resolution of UTC_NOW the same across transaction (#4248)
## Relevant issue(s)

Resolves #4231 

## Description

This PR makes it so that across a transaction, all instances of
`UTC_NOW` will resolve to the same value.

The core work is done by adding a new field to `datastore.BasicTxn`,
called `ts` for timestamp. We also add function to the `datastore.Txn`
interface, `StartTS()` which returns the timestamp. This value is then
propogated down inside a context. The new `clock` package defines the
context key that is used for this purpose, and contains the functions
that add a timestamp to, and get a timestamp from, a given context.

You can see that many function signatures have been touched to
facilitate this passing down-and-through of the context:
`client.NewDocFromMap`, `client.NewDocWithID`, `doc.SetWithJSON`, etc.
In most of these places, the change will be straightforward, and the
reason for the change will hopefully be self-evident.

Additionally, this PR includes several integration tests which
illustrate the new functionality. Further, the client wrappers have all
been adjusted accordingly, such that the Txn wrappers fit the new
interface.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- WSL

---------

Co-authored-by: John-Alan Simmons <simmons.johnalan@gmail.com>