refactor: Remove p2p transactions (#4212)
## Relevant issue(s)

Resolves #4211

## Description

This PR removes as many transactions from the p2p system as possible to
allow for the LevelDB store to pass the integration tests.

This is part of a stacked PR and will be undrafted once I open the other
PRs.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Integration tests

Specify the platform(s) on which this was tested:
- MacOS